Use double precision for sum in top-p filtering
Ensure that even for very small `top_p` values or large
vocabularies, the cumulative sum remains accurate, avoiding or at
least reducing floating-point rounding errors.

+++ src/kernels/src/gpu_sampling.cu
@@ -234,10 +234,10 @@ __global__ void stageB_reduce_and_sample(
     // top-p within top-k
     int cutoff = K;
     if (top_p > 0.0f && top_p < 1.0f) {
-      float cum = 0.0f;
+      double cum = 0.0;
       for (int t = 0; t < K; ++t) {
-        cum += probs[t];
-        if (cum >= top_p) { cutoff = t + 1; break; }
+        cum += (double)probs[t];
+        if (cum >= (double)top_p) { cutoff = t + 1; break; }
       }
     }
